+++ source/isaaclab/isaaclab/envs/mdp/commands/pose_2d_command.py
@@ -65,7 +65,7 @@ class UniformPose2dCommand(CommandTerm):
         if self._track_success:
             self._succeeded = torch.zeros(self.num_envs, dtype=torch.bool, device=self.device)
 
-        self.cfg.cmd_hint = self.cfg.cmd_hint or "command/body/pose"
+        self.cfg.cmd_kind = self.cfg.cmd_kind or "command/body/pose"
         self.cfg.element_names = self.cfg.element_names or ["x", "y", "z", "heading"]
 
     def __str__(self) -> str:
