+++ source/isaaclab/isaaclab/envs/mdp/commands/pose_command.py
@@ -72,7 +72,7 @@ class UniformPoseCommand(CommandTerm):
         if self._track_success:
             self._succeeded = torch.zeros(self.num_envs, dtype=torch.bool, device=self.device)
 
-        self.cfg.cmd_hint = self.cfg.cmd_hint or "command/body/pose"
+        self.cfg.cmd_kind = self.cfg.cmd_kind or "command/body/pose"
         self.cfg.element_names = self.cfg.element_names or ["x", "y", "z", "qw", "qx", "qy", "qz"]
 
     def __str__(self) -> str:
