+++ source/isaaclab/isaaclab/envs/mdp/commands/velocity_command.py
@@ -92,7 +92,7 @@ class UniformVelocityCommand(CommandTerm):
         self._error_yaw_sum = torch.zeros(self.num_envs, device=self.device)
         self._step_count = torch.zeros(self.num_envs, device=self.device)
 
-        self.cfg.cmd_hint = self.cfg.cmd_hint or "command/body/velocity"
+        self.cfg.cmd_kind = self.cfg.cmd_kind or "command/body/velocity"
         self.cfg.element_names = self.cfg.element_names or ["lin_vel_x", "lin_vel_y", "ang_vel_z"]
 
     def __str__(self) -> str:
