+++ source/isaaclab/isaaclab/managers/manager_term_cfg.py
@@ -118,7 +118,7 @@ class CommandTermCfg:
     debug_vis: bool = False
     """Whether to visualize debug information. Defaults to False."""
 
-    cmd_hint: str | None = None  # type hint for the command for deployment
+    cmd_kind: str | None = None  # type hint for the command for deployment
     element_names: list[str] | list[list[str]] | None = None  # element names for the command for deployment
