+++ source/isaaclab/isaaclab/sensors/contact_sensor/base_contact_sensor_data.py
@@ -12,7 +12,7 @@ from abc import ABC, abstractmethod
 import warp as wp
 from leapp import InputKindEnum
 
-from isaaclab.utils.leapp_semantics import (
+from isaaclab.utils.leapp.leapp_semantics import (
     POSE7_ELEMENT_NAMES,
     QUAT_WXYZ_ELEMENT_NAMES,
     XYZ_ELEMENT_NAMES,
