+++ source/isaaclab/isaaclab/sensors/imu/base_imu_data.py
@@ -12,7 +12,7 @@ from abc import ABC, abstractmethod
 import warp as wp
 from leapp import InputKindEnum
 
-from isaaclab.utils.leapp_semantics import XYZ_ELEMENT_NAMES, leapp_tensor_semantics
+from isaaclab.utils.leapp.leapp_semantics import XYZ_ELEMENT_NAMES, leapp_tensor_semantics
 from isaaclab.utils.warp import ProxyArray
