+++ source/isaaclab/isaaclab/sensors/ray_caster/ray_caster_data.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 import warp as wp
 
 from isaaclab.utils.warp import ProxyArray
-from isaaclab.utils.leapp_semantics import leapp_tensor_semantics
+from isaaclab.utils.leapp.leapp_semantics import leapp_tensor_semantics
 
 class RayCasterData:
     """Data container for the ray-cast sensor.
