+++ source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/dexsuite/mdp/commands/pose_commands.py
@@ -82,7 +82,7 @@ class ObjectUniformPoseCommand(CommandTerm):
         self.success_visualizer = VisualizationMarkers(self.cfg.success_visualizer_cfg)
         self.success_visualizer.set_visibility(True)
 
-        self.cfg.cmd_hint = self.cfg.cmd_hint or "command/body/pose"
+        self.cfg.cmd_kind = self.cfg.cmd_kind or "command/body/pose"
         self.cfg.element_names = self.cfg.element_names or ["x", "y", "z", "qw", "qx", "qy", "qz"]
 
     def __str__(self) -> str:
